backend/requirements.txt
# Core
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication
python-jose==3.3.0
passlib==1.7.4
python-dotenv==1.0.0

# Data Validation
pydantic==2.5.2
pydantic[email]==2.5.2
email-validator==2.1.0

# AI/ML
openai==1.3.9
langchain==0.1.7
faiss-cpu==1.7.4
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Utilities
requests==2.31.0
redis==5.0.1
